Move Ray and intersection methods to helper module
Also move tuv_to_xyz method to math.rs
