+++ common/src/character_controller/collision.rs
@@ -2,7 +2,7 @@
 
 use tracing::error;
 
-use crate::{graph::Graph, graph_collision, math, proto::Position};
+use crate::{collision_math::Ray, graph::Graph, graph_collision, math, proto::Position};
 
 /// Checks for collisions when a character moves with a character-relative displacement vector of `relative_displacement`.
 pub fn check_collision(
@@ -22,7 +22,7 @@ pub fn check_collision(
     let displacement_norm = displacement_sqr.sqrt();
     let displacement_normalized = relative_displacement / displacement_norm;
 
-    let ray = graph_collision::Ray::new(math::origin(), displacement_normalized);
+    let ray = Ray::new(math::origin(), displacement_normalized);
     let tanh_distance = displacement_norm.tanh();
 
     let cast_hit = graph_collision::sphere_cast(
