+++ common/src/chunk_collision.rs
@@ -1,5 +1,5 @@
 use crate::{
-    graph_collision::Ray,
+    collision_math::Ray,
     math,
     node::{ChunkLayout, Coords, VoxelData},
     world::Material,
@@ -92,13 +92,13 @@ fn find_face_collision(
     for t in bounding_box.grid_planes(t_axis) {
         // Find a normal to the grid plane. Note that (t, 0, 0, x) is a normal of the plane whose closest point
         // to the origin is (x, 0, 0, t), and we use that fact here.
-        let normal = math::lorentz_normalize(&tuv_to_xyz(
+        let normal = math::lorentz_normalize(&math::tuv_to_xyz(
             t_axis,
             na::Vector4::new(1.0, 0.0, 0.0, layout.grid_to_dual(t)),
         ));
 
         let Some(new_tanh_distance) =
-            solve_sphere_plane_intersection(ray, &normal, collider_radius.sinh())
+            ray.solve_sphere_plane_intersection(&normal, collider_radius.sinh())
         else {
             continue;
         };
@@ -138,7 +138,7 @@ fn find_face_collision(
         if !voxel_is_solid(
             voxel_data,
             layout,
-            tuv_to_xyz(t_axis, [voxel_t, voxel_u, voxel_v]),
+            math::tuv_to_xyz(t_axis, [voxel_t, voxel_u, voxel_v]),
         ) {
             continue;
         }
@@ -171,12 +171,12 @@ fn find_edge_collision(
     // Loop through all grid lines overlapping the bounding box
     for (u, v) in bounding_box.grid_lines(u_axis, v_axis) {
         // Compute vectors Lorentz-orthogonal to the edge and to each other
-        let edge_normal0 = math::lorentz_normalize(&tuv_to_xyz(
+        let edge_normal0 = math::lorentz_normalize(&math::tuv_to_xyz(
             t_axis,
             na::Vector4::new(0.0, 1.0, 0.0, layout.grid_to_dual(u)),
         ));
 
-        let edge_normal1 = tuv_to_xyz(
+        let edge_normal1 = math::tuv_to_xyz(
             t_axis,
             na::Vector4::new(0.0, 0.0, 1.0, layout.grid_to_dual(v)),
         );
@@ -184,8 +184,7 @@ fn find_edge_collision(
             &(edge_normal1 - edge_normal0 * math::mip(&edge_normal0, &edge_normal1)),
         );
 
-        let Some(new_tanh_distance) = solve_sphere_line_intersection(
-            ray,
+        let Some(new_tanh_distance) = ray.solve_sphere_line_intersection(
             &edge_normal0,
             &edge_normal1,
             collider_radius.sinh(),
@@ -214,7 +213,7 @@ fn find_edge_collision(
                 !voxel_is_solid(
                     voxel_data,
                     layout,
-                    tuv_to_xyz(t_axis, [voxel_t, voxel_u, voxel_v]),
+                    math::tuv_to_xyz(t_axis, [voxel_t, voxel_u, voxel_v]),
                 )
             })
         }) {
@@ -271,8 +270,7 @@ fn find_vertex_collision(
                 - vertex_normal1 * math::mip(&vertex_normal1, &vertex_normal2)),
         );
 
-        let Some(new_tanh_distance) = solve_sphere_point_intersection(
-            ray,
+        let Some(new_tanh_distance) = ray.solve_sphere_point_intersection(
             &vertex_normal0,
             &vertex_normal1,
             &vertex_normal2,
@@ -303,117 +301,6 @@ fn find_vertex_collision(
     }
 
     hit
-}
-
-/// Finds the tanh of the distance a sphere will have to travel along a ray before it
-/// intersects the given plane.
-fn solve_sphere_plane_intersection(
-    ray: &Ray,
-    plane_normal: &na::Vector4<f32>,
-    sinh_radius: f32,
-) -> Option<f32> {
-    let mip_pos_a = math::mip(&ray.position, plane_normal);
-    let mip_dir_a = math::mip(&ray.direction, plane_normal);
-
-    solve_quadratic(
-        mip_pos_a.powi(2) - sinh_radius.powi(2),
-        mip_pos_a * mip_dir_a,
-        mip_dir_a.powi(2) + sinh_radius.powi(2),
-    )
-}
-
-/// Finds the tanh of the distance a sphere will have to travel along a ray before it
-/// intersects the given line.
-fn solve_sphere_line_intersection(
-    ray: &Ray,
-    line_normal0: &na::Vector4<f32>,
-    line_normal1: &na::Vector4<f32>,
-    sinh_radius: f32,
-) -> Option<f32> {
-    let mip_pos_a = math::mip(&ray.position, line_normal0);
-    let mip_dir_a = math::mip(&ray.direction, line_normal0);
-    let mip_pos_b = math::mip(&ray.position, line_normal1);
-    let mip_dir_b = math::mip(&ray.direction, line_normal1);
-
-    solve_quadratic(
-        mip_pos_a.powi(2) + mip_pos_b.powi(2) - sinh_radius.powi(2),
-        mip_pos_a * mip_dir_a + mip_pos_b * mip_dir_b,
-        mip_dir_a.powi(2) + mip_dir_b.powi(2) + sinh_radius.powi(2),
-    )
-}
-
-/// Finds the tanh of the distance a sphere will have to travel along a ray before it
-/// intersects the given point.
-fn solve_sphere_point_intersection(
-    ray: &Ray,
-    point_normal0: &na::Vector4<f32>,
-    point_normal1: &na::Vector4<f32>,
-    point_normal2: &na::Vector4<f32>,
-    sinh_radius: f32,
-) -> Option<f32> {
-    let mip_pos_a = math::mip(&ray.position, point_normal0);
-    let mip_dir_a = math::mip(&ray.direction, point_normal0);
-    let mip_pos_b = math::mip(&ray.position, point_normal1);
-    let mip_dir_b = math::mip(&ray.direction, point_normal1);
-    let mip_pos_c = math::mip(&ray.position, point_normal2);
-    let mip_dir_c = math::mip(&ray.direction, point_normal2);
-
-    solve_quadratic(
-        mip_pos_a.powi(2) + mip_pos_b.powi(2) + mip_pos_c.powi(2) - sinh_radius.powi(2),
-        mip_pos_a * mip_dir_a + mip_pos_b * mip_dir_b + mip_pos_c * mip_dir_c,
-        mip_dir_a.powi(2) + mip_dir_b.powi(2) + mip_dir_c.powi(2) + sinh_radius.powi(2),
-    )
-}
-
-/// Finds the lower solution `x` of `constant_term + 2 * half_linear_term * x + quadratic_term * x * x == 0`
-/// if such a solution exists and is non-negative. Assumes that `quadratic_term` is positive. Double-roots are
-/// ignored.
-///
-/// If the lower solution is negative, but a small perturbation to the constant term would make it 0, this function
-/// returns 0.
-fn solve_quadratic(constant_term: f32, half_linear_term: f32, quadratic_term: f32) -> Option<f32> {
-    const EPSILON: f32 = 1e-4;
-
-    // If the linear term is positive, the lower solution is negative, and we're not interested. If the
-    // linear term is zero, the solution can only be non-negative if the constant term is also zero,
-    // which results in a double-root, which we also ignore.
-    if half_linear_term >= 0.0 {
-        return None;
-    }
-
-    // If the constant term is negative, the lower solution must also be negative. To avoid precision issues
-    // allowing a collider to clip through a surface, we treat small negative constant terms as zero, which
-    // results in a lower solution of zero.
-    if constant_term <= 0.0 {
-        return if constant_term > -EPSILON {
-            Some(0.0)
-        } else {
-            None
-        };
-    }
-
-    let discriminant = half_linear_term * half_linear_term - quadratic_term * constant_term;
-    if discriminant <= 0.0 {
-        return None;
-    }
-
-    // We use an alternative quadratic formula to ensure that we return a positive number if `constant_term > 0.0`.
-    // Otherwise, the edge case of a small positive `constant_term` could be mishandled.
-    // The denominator cannot be zero because both of its terms are positive.
-    Some(constant_term / (-half_linear_term + discriminant.sqrt()))
-}
-
-/// Converts from t-u-v coordinates to x-y-z coordinates. t-u-v coordinates are a permuted version of x-y-z coordinates.
-/// `t_axis` determines which of the three x-y-z coordinates corresponds to the t-coordinate. This function works with
-/// any indexable entity with at least three entries. Any entry after the third entry is ignored.
-fn tuv_to_xyz<T: std::ops::IndexMut<usize, Output = N>, N: Copy>(t_axis: usize, tuv: T) -> T {
-    let mut result = tuv;
-    (
-        result[t_axis],
-        result[(t_axis + 1) % 3],
-        result[(t_axis + 2) % 3],
-    ) = (result[0], result[1], result[2]);
-    result
 }
 
 /// Checks whether a voxel can be collided with. Any non-void voxel falls under this category.
@@ -503,7 +390,6 @@ mod tests {
     use crate::node::VoxelData;
 
     use super::*;
-    use approx::*;
 
     /// Helper structure used to reduce the number of parameters to pass around with tests
     struct TestSphereCastContext {
@@ -834,264 +720,6 @@ mod tests {
                 assert!(chunk_sphere_cast_wrapper(&ctx, ray, tanh_distance).is_none());
             },
         )
-    }
-
-    #[test]
-    fn solve_sphere_plane_intersection_example() {
-        // Hit the z=0 plane with a radius of 0.2
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(math::origin(), na::Vector4::new(0.8, 0.0, 0.6, 0.0));
-        let normal = -na::Vector4::z();
-        let hit_point = math::lorentz_normalize(
-            &ray.ray_point(solve_sphere_plane_intersection(&ray, &normal, 0.2_f32.sinh()).unwrap()),
-        );
-        assert_abs_diff_eq!(
-            math::mip(&hit_point, &normal),
-            0.2_f32.sinh(),
-            epsilon = 1e-4
-        );
-    }
-
-    #[test]
-    fn solve_sphere_plane_intersection_direct_hit() {
-        // Directly hit the z=0 plane with a ray 0.5 units away and a radius of 0.2.
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(math::origin(), na::Vector4::z());
-        let normal = -na::Vector4::z();
-        assert_abs_diff_eq!(
-            solve_sphere_plane_intersection(&ray, &normal, 0.2_f32.sinh()).unwrap(),
-            0.3_f32.tanh(),
-            epsilon = 1e-4
-        );
-    }
-
-    #[test]
-    fn solve_sphere_plane_intersection_miss() {
-        // No collision with the plane anywhere along the ray's line
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(math::origin(), na::Vector4::x());
-        let normal = -na::Vector4::z();
-        assert!(solve_sphere_plane_intersection(&ray, &normal, 0.2_f32.sinh()).is_none());
-    }
-
-    #[test]
-    fn solve_sphere_plane_intersection_margin() {
-        // Sphere is already contacting the plane, with some error
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.2))
-            * &Ray::new(math::origin(), na::Vector4::z());
-        let normal = -na::Vector4::z();
-        assert_eq!(
-            solve_sphere_plane_intersection(&ray, &normal, 0.2001_f32.sinh()).unwrap(),
-            0.0
-        );
-    }
-
-    #[test]
-    fn solve_sphere_line_intersection_example() {
-        // Hit the x=z=0 line with a radius of 0.2
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(
-                math::origin(),
-                na::Vector4::new(1.0, 2.0, 3.0, 0.0).normalize(),
-            );
-        let line_normal0 = na::Vector4::x();
-        let line_normal1 = na::Vector4::z();
-        let hit_point = math::lorentz_normalize(
-            &ray.ray_point(
-                solve_sphere_line_intersection(&ray, &line_normal0, &line_normal1, 0.2_f32.sinh())
-                    .unwrap(),
-            ),
-        );
-        // Measue the distance from hit_point to the line and ensure it's equal to the radius
-        assert_abs_diff_eq!(
-            (math::mip(&hit_point, &line_normal0).powi(2)
-                + math::mip(&hit_point, &line_normal1).powi(2))
-            .sqrt(),
-            0.2_f32.sinh(),
-            epsilon = 1e-4
-        );
-    }
-
-    #[test]
-    fn solve_sphere_line_intersection_direct_hit() {
-        // Directly hit the x=z=0 line with a ray 0.5 units away and a radius of 0.2.
-
-        // Ensure the ray is slightly off-center so that the distance math is shown to be correct
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.7, 0.0))
-            * math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(math::origin(), na::Vector4::z());
-        let line_normal0 = na::Vector4::x();
-        let line_normal1 = na::Vector4::z();
-        assert_abs_diff_eq!(
-            solve_sphere_line_intersection(&ray, &line_normal0, &line_normal1, 0.2_f32.sinh())
-                .unwrap(),
-            0.3_f32.tanh(),
-            epsilon = 1e-4
-        );
-    }
-
-    #[test]
-    fn solve_sphere_line_intersection_miss() {
-        // No collision with the line anywhere along the ray's line
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(math::origin(), na::Vector4::x());
-        let line_normal0 = na::Vector4::x();
-        let line_normal1 = na::Vector4::z();
-        assert!(
-            solve_sphere_line_intersection(&ray, &line_normal0, &line_normal1, 0.2_f32.sinh())
-                .is_none()
-        );
-    }
-
-    #[test]
-    fn solve_sphere_line_intersection_margin() {
-        // Sphere is already contacting the line, with some error
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.2))
-            * &Ray::new(math::origin(), na::Vector4::z());
-        let line_normal0 = na::Vector4::x();
-        let line_normal1 = na::Vector4::z();
-        assert_eq!(
-            solve_sphere_line_intersection(&ray, &line_normal0, &line_normal1, 0.2001_f32.sinh())
-                .unwrap(),
-            0.0
-        );
-    }
-
-    #[test]
-    fn solve_sphere_line_intersection_precision() {
-        // Using ray coordinates determined empirically from manual playtesting, show that the
-        // current implementation of `solve_sphere_line_intersection` provides better results
-        // than an arguably-simpler implementation involving the line's position and direction.
-        // Similar reasoning can also apply to `solve_sphere_point_intersection` even though it is
-        // not tested explicitly in the same way.
-        let ray = Ray::new(
-            na::Vector4::new(-0.019093871, -0.0014823675, 0.059645057, 1.0019588),
-            na::Vector4::new(-0.02954007, 0.9965602, 0.07752046, 0.003702946),
-        );
-        let line_normal0 = na::Vector4::<f32>::x();
-        let line_normal1 = na::Vector4::<f32>::y();
-        let radius = 0.019090926_f32;
-        // The following returns wrong results in the other implementation, so we test this case
-        // to make sure there are no regressions.
-        assert!(
-            solve_sphere_line_intersection(&ray, &line_normal0, &line_normal1, radius.sinh())
-                .is_none()
-        );
-    }
-
-    #[test]
-    fn solve_sphere_point_intersection_example() {
-        // Hit the origin with a radius of 0.2
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(
-                math::origin(),
-                na::Vector4::new(1.0, 2.0, 6.0, 0.0).normalize(),
-            );
-        let point_position = math::origin();
-        let point_normal0 = na::Vector4::x();
-        let point_normal1 = na::Vector4::y();
-        let point_normal2 = na::Vector4::z();
-        let hit_point = math::lorentz_normalize(
-            &ray.ray_point(
-                solve_sphere_point_intersection(
-                    &ray,
-                    &point_normal0,
-                    &point_normal1,
-                    &point_normal2,
-                    0.2_f32.sinh(),
-                )
-                .unwrap(),
-            ),
-        );
-        assert_abs_diff_eq!(
-            -math::mip(&hit_point, &point_position),
-            0.2_f32.cosh(),
-            epsilon = 1e-4
-        );
-    }
-
-    #[test]
-    fn solve_sphere_point_intersection_direct_hit() {
-        // Directly hit the origin with a ray 0.5 units away and a radius of 0.2.
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(math::origin(), na::Vector4::z());
-        let point_normal0 = na::Vector4::x();
-        let point_normal1 = na::Vector4::y();
-        let point_normal2 = na::Vector4::z();
-        assert_abs_diff_eq!(
-            solve_sphere_point_intersection(
-                &ray,
-                &point_normal0,
-                &point_normal1,
-                &point_normal2,
-                0.2_f32.sinh()
-            )
-            .unwrap(),
-            0.3_f32.tanh(),
-            epsilon = 1e-4
-        );
-    }
-
-    #[test]
-    fn solve_sphere_point_intersection_miss() {
-        // No collision with the point anywhere along the ray's line
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.5))
-            * &Ray::new(math::origin(), na::Vector4::x());
-        let point_normal0 = na::Vector4::x();
-        let point_normal1 = na::Vector4::y();
-        let point_normal2 = na::Vector4::z();
-        assert!(solve_sphere_point_intersection(
-            &ray,
-            &point_normal0,
-            &point_normal1,
-            &point_normal2,
-            0.2_f32.sinh()
-        )
-        .is_none());
-    }
-
-    #[test]
-    fn solve_sphere_point_intersection_margin() {
-        // Sphere is already contacting the point, with some error
-        let ray = math::translate_along(&na::Vector3::new(0.0, 0.0, -0.2))
-            * &Ray::new(math::origin(), na::Vector4::z());
-        let point_normal0 = na::Vector4::x();
-        let point_normal1 = na::Vector4::y();
-        let point_normal2 = na::Vector4::z();
-        assert_eq!(
-            solve_sphere_point_intersection(
-                &ray,
-                &point_normal0,
-                &point_normal1,
-                &point_normal2,
-                0.2001_f32.sinh()
-            )
-            .unwrap(),
-            0.0
-        );
-    }
-
-    #[test]
-    fn solve_quadratic_example() {
-        let a = 1.0;
-        let b = -2.0;
-        let c = 0.2;
-        let x = solve_quadratic(c, b / 2.0, a).unwrap();
-
-        // x should be a solution
-        assert_abs_diff_eq!(a * x * x + b * x + c, 0.0, epsilon = 1e-4);
-
-        // x should be the smallest solution, less than the parabola's vertex.
-        assert!(x < -b / (2.0 * a));
-    }
-
-    #[test]
-    fn tuv_to_xyz_example() {
-        assert_eq!(tuv_to_xyz(0, [2, 4, 6]), [2, 4, 6]);
-        assert_eq!(tuv_to_xyz(1, [2, 4, 6]), [6, 2, 4]);
-        assert_eq!(tuv_to_xyz(2, [2, 4, 6]), [4, 6, 2]);
-
-        assert_eq!(tuv_to_xyz(1, [2, 4, 6, 8]), [6, 2, 4, 8]);
     }
 
     /// Any voxel AABB should at least cover a capsule-shaped region consisting of all points
