+++ common/src/lib.rs
@@ -13,6 +13,7 @@ pub mod character_controller;
 pub mod chunk_collision;
 mod chunks;
 pub mod codec;
+pub mod collision_math;
 pub mod cursor;
 pub mod dodeca;
 pub mod graph;
