+++ common/src/math.rs
@@ -166,6 +166,19 @@ pub fn rotation_between_axis<N: RealField + Copy>(
     ))
 }
 
+/// Converts from t-u-v coordinates to x-y-z coordinates. t-u-v coordinates are a permuted version of x-y-z coordinates.
+/// `t_axis` determines which of the three x-y-z coordinates corresponds to the t-coordinate. This function works with
+/// any indexable entity with at least three entries. Any entry after the third entry is ignored.
+pub fn tuv_to_xyz<T: std::ops::IndexMut<usize, Output = N>, N: Copy>(t_axis: usize, tuv: T) -> T {
+    let mut result = tuv;
+    (
+        result[t_axis],
+        result[(t_axis + 1) % 3],
+        result[(t_axis + 2) % 3],
+    ) = (result[0], result[1], result[2]);
+    result
+}
+
 fn minkowski_outer_product<N: RealField + Copy>(
     a: &na::Vector4<N>,
     b: &na::Vector4<N>,
@@ -331,5 +344,14 @@ mod tests {
         let expected = na::UnitQuaternion::rotation_between_axis(&from, &to).unwrap();
         let actual = rotation_between_axis(&from, &to, 1e-5).unwrap();
         assert_abs_diff_eq!(expected, actual, epsilon = 1.0e-5);
+    }
+
+    #[test]
+    fn tuv_to_xyz_example() {
+        assert_eq!(tuv_to_xyz(0, [2, 4, 6]), [2, 4, 6]);
+        assert_eq!(tuv_to_xyz(1, [2, 4, 6]), [6, 2, 4]);
+        assert_eq!(tuv_to_xyz(2, [2, 4, 6]), [4, 6, 2]);
+
+        assert_eq!(tuv_to_xyz(1, [2, 4, 6, 8]), [6, 2, 4, 8]);
     }
 }
